feat: add select with options groups
